fixed it yay :)
